tests/requirements_test.txt
# Dependencias para testing del Sistema de Predicción de Tarifas
# Instalar con: pip install -r tests/requirements_test.txt

# Framework de testing principal
pytest==7.4.3
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-xdist==3.3.1

# Testing de aplicaciones web
selenium==4.15.2
webdriver-manager==4.0.1

# Mocking y utilidades de testing
pytest-mock==3.12.0
responses==0.24.1
factory-boy==3.3.0

# Dependencias del proyecto (versiones mínimas)
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
geopandas>=0.14.0
folium>=0.15.0
plotly>=5.15.0
unidecode>=1.3.0

# Machine Learning para tests de predicciones
prophet>=1.1.4
statsmodels>=0.14.0
xgboost>=2.0.0
scikit-learn>=1.3.0

# Base de datos para tests de integración
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Utilidades adicionales para testing
freezegun==1.2.2  # Para tests con fechas
parameterized==0.9.0  # Tests paramétricos adicionales
testfixtures==7.2.2  # Fixtures avanzadas
faker==20.1.0  # Generación de datos de prueba

# Análisis de código y formateo
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Documentación de tests
pytest-clarity==1.0.1
pytest-sugar==0.9.7 